pyroscope
Module pyroscope

Structs

BasicAuth

PyroscopeAgent

PyroscopeAgent is the main object of the library. It is used to start and stop the profiler, schedule the timer, and send the profiler data to the server.

PyroscopeAgentBare

Marker struct for an Uninitialized state.

PyroscopeAgentBuilder

PyroscopeAgent Builder

PyroscopeAgentReady

Marker struct for a Ready state.

PyroscopeAgentRunning

Marker struct for a Running state.

PyroscopeConfig

Pyroscope Agent Configuration. This is the configuration that is passed to the agent.

Enums

Compression

ReportEncoding

Traits

PyroscopeAgentState

This trait is used to encode the state of the agent.

Functions

parse_http_headers_json

parse_vec_string_json