pyroscope::pyroscope
Struct PyroscopeAgent

pub struct PyroscopeAgent<S: PyroscopeAgentState> {
    pub backend: BackendImpl<BackendReady>,
    pub config: PyroscopeConfig,
    /* private fields */
}

PyroscopeAgent is the main object of the library. It is used to start and stop the profiler, schedule the timer, and send the profiler data to the server.

Fields
backend: BackendImpl<BackendReady>

Profiler backend

config: PyroscopeConfig

Configuration Object

Implementations

impl<S: PyroscopeAgentState> PyroscopeAgent<S>

pub fn shutdown(self)

Properly shutdown the agent.

impl PyroscopeAgent<PyroscopeAgentBare>

pub fn builder<S: AsRef<str>>( url: S, application_name: S, ) -> PyroscopeAgentBuilder

Short-hand for PyroscopeAgentBuilder::new(url, application_name). This is a convenience method.

Example

let agent = PyroscopeAgent::builder("http://localhost:8080", "my-app").build()?;

pub fn default_builder() -> PyroscopeAgentBuilder

Short-hand for PyroscopeAgentBuilder::default(). This is a convenience method. Default URL is “http://localhost:4040”. Default application name is randomly generated.

Example

let agent = PyroscopeAgent::default_builder().build()?;

impl PyroscopeAgent<PyroscopeAgentReady>

pub fn start(self) -> Result<PyroscopeAgent<PyroscopeAgentRunning>>

Start profiling and sending data. The agent will keep running until stopped. The agent will send data to the server every 10s seconds.

Example

let agent = PyroscopeAgent::builder("http://localhost:8080", "my-app").build()?;
let agent_running = agent.start()?;

impl PyroscopeAgent<PyroscopeAgentRunning>

pub fn stop(self) -> Result<PyroscopeAgent<PyroscopeAgentReady>>

Stop the agent. The agent will stop profiling and send a last report to the server.

Example

let agent = PyroscopeAgent::builder("http://localhost:8080", "my-app").build()?;
let agent_running = agent.start()?;
// Expensive operation
let agent_ready = agent_running.stop();

pub fn tag_wrapper( &self, ) -> (impl Fn(String, String) -> Result<()>, impl Fn(String, String) -> Result<()>)

Return a tuple of functions to add and remove tags to the agent across thread boundaries. This function can be called multiple times.

Example

let agent = PyroscopeAgent::builder("http://localhost:8080", "my-app").build()?;
let agent_running = agent.start()?;
let (add_tag, remove_tag) = agent_running.tag_wrapper();

The functions can be later called from any thread.

Example

add_tag("key".to_string(), "value".to_string());
// some computation
remove_tag("key".to_string(), "value".to_string());

pub fn add_global_tag(&self, tag: Tag) -> Result<()>

Add a global Tag rule to the backend Ruleset. For tagging, it’s recommended to use the tag_wrapper function.

pub fn remove_global_tag(&self, tag: Tag) -> Result<()>

Remove a global Tag rule from the backend Ruleset. For tagging, it’s recommended to use the tag_wrapper function.

pub fn add_thread_tag(&self, thread_id: u64, tag: Tag) -> Result<()>

Add a thread Tag rule to the backend Ruleset. For tagging, it’s recommended to use the tag_wrapper function.

pub fn remove_thread_tag(&self, thread_id: u64, tag: Tag) -> Result<()>

Remove a thread Tag rule from the backend Ruleset. For tagging, it’s recommended to use the tag_wrapper function.