Struct pyroscope::pyroscope::PyroscopeAgentBuilder

pub struct PyroscopeAgentBuilder { /* private fields */ }

PyroscopeAgent Builder

Alternatively, you can use PyroscopeAgent::build() which is a short-hand for calling PyroscopeAgentBuilder::new()

Example

use pyroscope::pyroscope::PyroscopeAgentBuilder;
let builder = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app");
let agent = builder.build()?;

Implementations

impl PyroscopeAgentBuilder

pub fn new(url: impl AsRef<str>, application_name: impl AsRef<str>) -> Self

Create a new PyroscopeAgentBuilder object. url and application_name are required. tags and sample_rate are optional.

Example

let builder = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app");

pub fn url(self, url: impl AsRef<str>) -> Self

Set the Pyroscope Server URL. This can be used if the Builder was initialized with the default trait. Default is “http://localhost:4040”.

Example

let builder = PyroscopeAgentBuilder::default()
.url("http://localhost:8080")
.build()?;

pub fn application_name(self, application_name: impl AsRef<str>) -> Self

Set the Application Name. This can be used if the Builder was initialized with the default trait. Default is a randomly generated name.

Example

let builder = PyroscopeAgentBuilder::default()
.application_name("my-app")
.build()?;

pub fn backend(self, backend: BackendImpl<BackendUninitialized>) -> Self

Set the agent backend. Default is void-backend.

Example

let builder = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app")
.backend(PprofConfig::new().sample_rate(100))
.build()?;

pub fn auth_token(self, auth_token: impl AsRef<str>) -> Self

Set JWT authentication token. This is optional. If not set, the agent will not send any authentication token.

#Example

let builder = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app")
.auth_token("my-token")
.build()
?;

pub fn basic_auth( self, username: impl AsRef<str>, password: impl AsRef<str>, ) -> Self

pub fn func(self, func: fn(_: Report) -> Report) -> Self

Set the Function. This is optional. If not set, the agent will not apply any function. #Example

let builder = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app")
.func(|report| {
   report
   })
   .build()
   ?;

pub fn tags(self, tags: Vec<(&str, &str)>) -> Self

Set tags. Default is empty.

Example

let builder = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app")
.tags(vec![("env", "dev")])
.build()?;

pub fn compression(self, compression: Compression) -> Self

Set the http request body compression.

Example

use pyroscope::pyroscope::PyroscopeConfig;
let agent = PyroscopeAgentBuilder::new("http://localhost:8080", "my-app")
    .compression(GZIP)
    .build();

pub fn report_encoding(self, report_encoding: ReportEncoding) -> Self

pub fn tenant_id(self, tenant_id: String) -> Self

pub fn http_headers(self, http_headers: HashMap<String, String>) -> Self

pub fn build(self) -> Result<PyroscopeAgent<PyroscopeAgentReady>>

Initialize the backend, timer and return a PyroscopeAgent with Ready state. While you can call this method, you should call it through the PyroscopeAgent.build() method.