Struct pyroscope::pyroscope::PyroscopeConfig

pub struct PyroscopeConfig {
    pub url: String,
    pub application_name: String,
    pub tags: HashMap<String, String>,
    pub sample_rate: u32,
    pub spy_name: String,
    pub auth_token: Option<String>,
    pub basic_auth: Option<BasicAuth>,
    pub func: Option<fn(_: Report) -> Report>,
    pub compression: Option<Compression>,
    pub report_encoding: ReportEncoding,
    pub tenant_id: Option<String>,
    pub http_headers: HashMap<String, String>,
}

Pyroscope Agent Configuration. This is the configuration that is passed to the agent.

Example

use pyroscope::pyroscope::PyroscopeConfig;
let config = PyroscopeConfig::new("http://localhost:8080", "my-app");

Fields
url: String

Pyroscope Server Address

application_name: String

Application Name

tags: HashMap<String, String>

Tags

sample_rate: u32

Sample Rate

spy_name: String

Spy Name

auth_token: Option<String>

Authentication Token

basic_auth: Option<BasicAuth>func: Option<fn(_: Report) -> Report>

Function to apply

compression: Option<Compression>

Pyroscope http request body compression

report_encoding: ReportEncodingtenant_id: Option<String>http_headers: HashMap<String, String>
Implementations

impl PyroscopeConfig

pub fn new(url: impl AsRef<str>, application_name: impl AsRef<str>) -> Self

Create a new PyroscopeConfig object. url and application_name are required. tags and sample_rate are optional. If sample_rate is not specified, it will default to 100.

Example

let config = PyroscopeConfig::new("http://localhost:8080", "my-app");

pub fn url(self, url: impl AsRef<str>) -> Self

pub fn application_name(self, application_name: impl AsRef<str>) -> Self

pub fn sample_rate(self, sample_rate: u32) -> Self

Set the Sample rate.

pub fn spy_name(self, spy_name: String) -> Self

Set the Spy Name.

pub fn auth_token(self, auth_token: String) -> Self

pub fn basic_auth(self, username: String, password: String) -> Self

pub fn func(self, func: fn(_: Report) -> Report) -> Self

Set the Function.

pub fn tags(self, tags: Vec<(&str, &str)>) -> Self

Set the tags.

Example

use pyroscope::pyroscope::PyroscopeConfig;
let config = PyroscopeConfig::new("http://localhost:8080", "my-app")
   .tags(vec![("env", "dev")])?;

pub fn compression(self, compression: Compression) -> Self

Set the http request body compression.

Example

use pyroscope::pyroscope::PyroscopeConfig;
let config = PyroscopeConfig::new("http://localhost:8080", "my-app")
    .compression(GZIP);

pub fn report_encoding(self, report_encoding: ReportEncoding) -> Self

pub fn tenant_id(self, tenant_id: String) -> Self

pub fn http_headers(self, http_headers: HashMap<String, String>) -> Self